Refactor drop tables cases to a shared file/logic
